ADD Refactor to extract NGSI Behavior to diferent file
