+++ lib/errors.js
@@ -55,6 +55,10 @@ module.exports = {
     TypeNotFound: function(id, type) {
         this.name = 'TYPE_NOT_FOUND';
         this.message = 'Type : ' + type + ' not found for device with id: ' + id;
+    },
+    BadRequestNorth: function(msg) {
+        this.name = 'BAD_REQUEST_NORTH';
+        this.message = 'The request was not well formed:' + msg;
     }
 
 };
