+++ lib/services/contextServer.js
@@ -204,19 +204,18 @@ function handleNotification(req, res, next) {
     }
 }
 
-function loadContextRoutes(app) {
-    app.post('/NGSI10/updateContext', handleUpdate);
-    app.post('/NGSI10/queryContext', handleQuery);
-    app.post('/notification', handleNotification);
-}
-
-
 function setUpdateHandler(newHandler) {
     updateHandler = newHandler;
 }
 
 function setQueryHandler(newHandler) {
     queryHandler = newHandler;
+}
+
+function loadContextRoutes(app) {
+    app.post('/NGSI10/updateContext', handleUpdate);
+    app.post('/NGSI10/queryContext', handleQuery);
+    app.post('/notification', handleNotification);
 }
 
 exports.loadContextRoutes = loadContextRoutes;
