+++ lib/services/northboundServer.js
@@ -26,6 +26,7 @@ var http = require('http'),
     express = require('express'),
     northboundServer,
     contextServer = require('./contextServer'),
+    deviceProvisioning = require('./deviceProvisioningServer'),
     logger = require('fiware-node-logger');
 
 function handleError(error, req, res, next) {
@@ -63,6 +64,7 @@ function start(config, callback) {
     }
 
     contextServer.loadContextRoutes(northboundServer.app);
+    deviceProvisioning.loadContextRoutes(northboundServer.app);
 
     northboundServer.app.use(handleError);
